feature: allow renaming of folders, allow removing of pages without their attachments
- Renaming of folders was brought up in #149.
- Deleting pages without deleting their subpages (subpages are
  technically just attachments) was brought up in #150.

+++ otterwiki/wiki.py
@@ -932,13 +932,21 @@ class Page:
             ),
         )
 
-    def delete(self, message, author):
+    def delete(self, message, author, recursive=True):
         if not has_permission("WRITE"):
             abort(403)
         if empty(message):
             message = "{} deleted.".format(self.pagename)
+        files = []
+        if self.exists:
+            files.append(self.filename)
+        if recursive:
+            files.append(self.attachment_directoryname)
+        if len(files) < 1:
+            toast("Nothing to delete.")
+            return redirect(url_for("view", path=self.pagepath))
         storage.delete(
-            [self.filename, self.attachment_directoryname],
+            files,
             message=message,
             author=author,
         )
